# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a CRUD application.",
        "title": "Boolang",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Lawrence Onaulogho",
            "url": "https://github.com/asekhamhe/",
            "email": "lawrence[at][gmail][dot][com]"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/books": {
            "get": {
                "description": "fetch list of books",
                "produces": [
                    "application/json"
                ],
                "summary": "GetBooks",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new book",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "CreateBook",
                "parameters": [
                    {
                        "description": "book model",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateBook"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.CreateBook"
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "description": "fetch a single book",
                "produces": [
                    "application/json"
                ],
                "summary": "GetBook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.Book"
                        }
                    }
                }
            }
        },
        "/home": {
            "get": {
                "description": "Test connection",
                "produces": [
                    "text/plain"
                ],
                "summary": "HomePage",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Book": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "string"
                }
            }
        },
        "models.CreateBook": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "string"
                }
            }
        }
    }
}